First batch

+++ libc/src/ctype/toupper.cpp
@@ -16,7 +16,7 @@ namespace __llvm_libc {
 // TODO: Currently restricted to default locale.
 // These should be extended using locale information.
 LLVM_LIBC_FUNCTION(int, toupper, (int c)) {
-  if (internal::islower(c))
+  if (internal::islower(static_cast<unsigned>(c)))
     return c + 'A' - 'a';
   return c;
 }
